undo util move
