+++ pkg/destinations/duckdb/insert.go
@@ -2,7 +2,7 @@ package duckdb
 
 import (
 	"fmt"
-	"github.com/scratchdata/scratchdata/pkg/util"
+	"github.com/scratchdata/scratchdata/util"
 	"os"
 	"path/filepath"
