+++ pkg/destinations/redshift/insert.go
@@ -2,7 +2,7 @@ package redshift
 
 import (
 	"fmt"
-	"github.com/scratchdata/scratchdata/pkg/util"
+	"github.com/scratchdata/scratchdata/util"
 	"path/filepath"
 
 	"os"
