+++ pkg/storage/queue/sqs/sqs.go
@@ -3,7 +3,7 @@ package sqs
 import (
 	"context"
 	"github.com/aws/aws-sdk-go-v2/service/sqs/types"
-	"github.com/scratchdata/scratchdata/pkg/util"
+	"github.com/scratchdata/scratchdata/util"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
